refactor(slideModal): 이벤트 관련 내용 컴포넌트 분리 - ModalHandle 컴포넌트에 드래그 상태 관리 및 터치 이벤트 핸들러 분리
